refactor(1159): refactor form validation schema to be reusable, refs: #1159
