+++ frontend/src/components/grantaccess/GrantDelegatedAdmin.vue
@@ -1,17 +1,14 @@
 <script setup lang="ts">
 import { ref, type PropType } from 'vue';
 import router from '@/router';
-import { number, object, string } from 'yup';
 import { ErrorMessage, Field, Form as VeeForm } from 'vee-validate';
-
-import { UserType } from 'fam-app-acsctl-api';
-
 import Dropdown from 'primevue/dropdown';
 import Button from '@/components/common/Button.vue';
 import { IconSize } from '@/enum/IconEnum';
-
 import { isLoading } from '@/store/LoadingState';
+import { UserType } from 'fam-app-acsctl-api';
 import type { FamRoleDto } from 'fam-admin-mgmt-api/model';
+import { formValidationSchema } from '@/services/utils';
 
 const props = defineProps({
     delegatedRoleOptions: {
@@ -27,27 +24,6 @@ const defaultFormData = {
     roleId: null as number | null,
 };
 const formData = ref(JSON.parse(JSON.stringify(defaultFormData))); // clone default input
-const formValidationSchema = object({
-    userId: string()
-        .required('User ID is required')
-        .min(2, 'User ID must be at least 2 characters')
-        .nullable(),
-    roleId: number().required('Please select a value'),
-    forestClientNumbers: string()
-        .when('roleId', {
-            is: (_role_id: number) => isAbstractRoleSelected(),
-            then: () =>
-                string()
-                    .nullable()
-                    .transform((curr, orig) => (orig === '' ? null : curr)) // Accept either null or value
-                    .matches(/^[0-9,\b]+$/, 'Please enter a digit or comma')
-                    .matches(
-                        /^\d{8}(,?\d{8})*$/,
-                        'Please enter a Forest Client ID with 8 digits long'
-                    ),
-        })
-        .nullable(),
-});
 
 /* ------------------ User information method ------------------------- */
 const userDomainChange = (selectedDomain: string) => {
@@ -124,7 +100,7 @@ const handleSubmit = async () => {
     <VeeForm
         ref="form"
         v-slot="{ errors, meta }"
-        :validation-schema="formValidationSchema"
+        :validation-schema="formValidationSchema(isAbstractRoleSelected())"
         as="div"
     >
         <div class="page-body">
